modify watershed scripts to use 64 bit unsingned integers
Former-commit-id: c28b68eebd1c94ddcecd6f7566f75bdd0ac51c93
Former-commit-id: 2a76d00b988e1adfdd7fdc00de70cd32207b6e94

+++ pipeline/scripts/watershed_merge.py
@@ -16,8 +16,8 @@ dendValues = numpy.fromfile('../watershed/data/input.dend_values', dtype='float3
 dendValues_dset = merged_file.create_dataset('/dendValues', data=dendValues, dtype='float32')
 
 
-dend = numpy.fromfile('../watershed/data/input.dend_pairs', dtype = 'uint32').reshape(-1,2).transpose()
-dend_dset = merged_file.create_dataset('/dend', data=dend, dtype='uint32')
+dend = numpy.fromfile('../watershed/data/input.dend_pairs', dtype = 'uint64').reshape(-1,2).transpose()
+dend_dset = merged_file.create_dataset('/dend', data=dend, dtype='uint64')
 
 #Read the metadata to find out how many chunks we have
 metadata = numpy.fromfile('../watershed/data/input.metadata', dtype='uint32')[2:5]
@@ -39,7 +39,7 @@ for z_chunk in range(metadata[0]):
 
             print chunk_size
 
-            main_chunk = numpy.fromfile('../watershed/data/input.chunks/{0}/{1}/{2}/.seg'.format(x_chunk, y_chunk, z_chunk), dtype='uint32').reshape(chunk_size)
+            main_chunk = numpy.fromfile('../watershed/data/input.chunks/{0}/{1}/{2}/.seg'.format(x_chunk, y_chunk, z_chunk), dtype='uint64').reshape(chunk_size)
             main_dset[zabs:zabs+chunk_size[0], yabs:yabs+chunk_size[1], xabs:xabs+chunk_size[2]] = main_chunk
